test(domain): move the collinearity screen's tests to a sibling file
Inline, they made it the crate's largest file at 733 lines; the module itself
is mid-pack at 373. Every comparable module here already splits them out.
